Variable string (#60)
* Remove need for string length

Remove the "requestedLength" arguments and use allocatable strings now that these are available.

* Update the documentation

The readme is updated. Add users guide in markdown format and remove latex version.

* Add more allocatable strings

Remove more dependences on fixed length character variables.

* Update UsersGuide.md

Add links to examples. Add info about optData functionality to the documentation.

* bump minimal CMake version

* Update UsersGuide.md

Align code blocks.

* Update UsersGuide.md

More code formatting. Fix link to Apple.

* Update UsersGuide.md

Caught a couple more alignment errors.

* Update UsersGuide.md

Fix an escape for a single underscore.

* Update UsersGuide.md

Fix code examples formatting.

* Update UsersGuide.md

Fix code example object names.

* AddLegacyCall

Add back the legacy calls to stringValue and stringValueForKey so that the changes in this version are non-breaking. Those functions are now generic with either the requested length or not. Going forward, one should use the versions without the requested length and use allocated strings.

* CharacterTruncation

Change the returned character length in the legacy function to what was originally there to get rid of the string truncation warning issue.

* Increase coverage

Increase coverage in stringValueForKey

* Logical return

Fix logical return and add test for it.

* Update ValueClassTests.f90

cover more stringValue tests

* Add a News.md section

A news page hanging off of the readme page is added for the latest and future news about changes.

* Update FTValueClass.f90

Just change the comments to reflect that stringValueA and stringValueR are overloaded as stringValue() and that the latter is deprecated but will still work with stringValue(requestedLength)

---------

Co-authored-by: Andrew Winters <andrew.ross.winters@liu.se>